Merging/copying changes from server
An initial submission to Open AIR 2015 was done on Nov 5-6. These
changes are merging in those changes from the server. There are a few
pieces of code that do need to be added back that never made it to the
server but are only needed for admin work, not guest viewing.